checkpoint, break things up into 2 steps
